+++ src/main/java/lt/codeacademy/learn/baigiamasis/security/WebSecurityConfig.java
@@ -39,8 +39,7 @@ public class WebSecurityConfig{
 		http
 			.csrf().disable()
 			.authorizeHttpRequests()
-//			.requestMatchers("/admin/**").hasAuthority(Roles.ADMIN)
-			.requestMatchers("/admin/**").permitAll()
+			.requestMatchers("/admin/**").hasAuthority(Roles.ADMIN)
 			.requestMatchers("/user/**").hasAuthority(Roles.USER)
 			.requestMatchers("/**").permitAll()
 			.anyRequest().authenticated()
